First change towards a two step signup
We create an (incomplete) account as soon as we get the DNSimple OAuth
dance results.

This change changes the signup flow so it matches what @aeden has done
for the Heroku connector. It also has the added benefit of not having to
fiddle with the session which has proved to be very hard to test.