Repeating the signup will update the DNSimple data
This is necessary to keep track of email changes or to allow access
tokens to be updated (which will happen when the user revokes the access
to the app but wants to use it again after some time).

These tests are being done at the Controller level but it would probably
make sense to test them from the actual domain model entry point (which
as of now is the Accounts module).