Update CNAME file if it exists
Since updating works differently than creating a new file we have to
handle the case specifically. The main difference is that you have to
provide the SHA of the current content in order to be able to update it.

In this commit we check for the CNAME file existence in the repo and
branch out to create it or update it, providing the necessary SHA hash
on later scenario.