Add spec for 1 click service behaviour
I've replaced the DnsimpleDummy with an Agent-based dummy. This way we
can alter the return values for stubbed methods and make very simple
assertions about the functions being called.

This is just an starting point to evaluate if it makes sense to have
cleverer dummies or if we should go for an actual mocking library.

I plan to apply the same pattern to the rest of the codebase in a
separate PR and see how it feels.