Improve default stubbed value
This is an easy enough way to make the default return value in the
dnsimple dummy a little bit cleverer for the create_record function.

It allows us to remove the ugly database update in the CNAME record
deletion test and makes the dummy behave a little bit more like the real
module.