Outline definition of TransactionalPipeline
More docs will come and a couple features have to be added and tested,
but this is enough to show progress and start replacing the existing
implementation of functions inside the Connections module to follow the
convention.

Both Connections.new_connection and Connections.remove_connection will
be changed to become transactional pipelines.