Adapt functions for the transactional pipeline
Adds the functions for adding and removing both ALIAS and CNAME records
complying with the requirements set by the TransactionalPipeline module.

The functions have been added to the Connections module keeping the old
ones in place until all the functions have been replaced. I also want to
be able to compare how the code looks inside and outside the pipeline.