Complementary function pipeline spike
This is an spike to test the grounds on a few things related to the
concerns that we discussed around the pipeline.

This commit introduces yet another version of the functions to add and
remove an alias record. However this new version bakes in the state
changes in the connection and how these are persisted. The change truly
makes one function the complementary of the other.

This is super helpful to define the rollback functions as you only have
to partially apply the complementary function and that's it. It can get
simpler than that.

On top of that I've written some tests for these functions to verify
how testable they are and I think we are safe on that end. There is
still some cleanup to do in the tests, but I consider them good enough
for now.

One final thing to mention is that this approach changes how the
pipelines will be structured: in the `new_connection` pipeline we were
persisting the `Connection` as the final step. However, with the new
approach, you would be creating a new empty connection as the first step
and update it along the way (this goes inline the concerns I expressed here:
https://github.com/dnsimple/addon-connector-ghpages/pull/17#issuecomment-260898243).

This first step would return a rollback function that deletes the
empty connection in case of any failure occurs during the pipeline
execution so I think the change is safe.

The `remove_connection` pipeline would remain the same as it makes sense
for the delete connection step to be the last one to be executed.