No callback for the last function of the pipeline
The main reason to do this is actually a limitation of the
`TransactionalPipeline` itself: the `_remove_connection` rollback
operation would create a new connection. The previous function's
callbacks would try to update a connection that no longer exists,
failing.

There is a workaround with actors that I'm already thinking on, but it's
complex enough to require a separate project.

I also consider the current solution more than enough as a first step
and I think it suits the needs I have perfectly.