Consider arguments in dummy stubbing
This change was prompted from an error that wasn't caught in a test (the
fix will come in the next commit). The problem was as simple as calling
the method with the parameters in the wrong order. The stubbing
supported in the dummies was super simple, and didn't consider the
argument values, so the test passed.

Since I wanted to pin down the error in a regression test before adding
the fix I've added the capability of stubbing the return value of a
dummy function depending on the parameter values. The required changes
make the agents a little more complicated, but I think it's worth it.